Updated Tuan03 use EJB CRUD product

+++ LeDonChung_Lab_Week03/src/main/java/vn/edu/iuh/fit/donchung/backend/repositories/ProductRepository.java
@@ -1,6 +1,5 @@
 package vn.edu.iuh.fit.donchung.backend.repositories;
 
-import jakarta.ejb.Local;
 import vn.edu.iuh.fit.donchung.backend.entities.Product;
 
 import java.util.List;
@@ -8,6 +7,6 @@ import java.util.Optional;
 
 public interface ProductRepository {
     List<Product> findAll();
-    Optional<Product> findById(Integer id);
+    Optional<Product> findById(Long id);
     Product save(Product product);
 }
